creating the initial logic for the GET endpoint

+++ src/main/java/com/recime/recipes/entity/recipe/repository/RecipeRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.recime.recipes.entity.recipe.model.Recipe;
 
 @Repository
-public interface RecipeRepository extends CrudRepository<Recipe, Long> {
+public interface RecipeRepository extends CrudRepository<Recipe, Integer> {
 
 }
